@@ -138,31 +140,31 @@
     "html-webpack-include-sibling-chunks-plugin": "^0.1.4",
     "html-webpack-plugin": "^3.2.0",
     "http-proxy-middleware": "^0.18.0",
-    "image-webpack-loader": "^4.2.0",
+    "image-webpack-loader": "~4.2.0",
     "lodash-webpack-plugin": "^0.11.5",
     "loud-rejection": "^1.6.0",
     "mini-css-extract-plugin": "^0.4.0",
     "minimist": "^1.2.0",
-    "mocha": "^5.1.1",
-    "nodemon": "^1.17.4",
-    "nyc": "^11.8.0",
+    "mocha": "^5.2.0",
+    "nodemon": "^1.17.5",
+    "nyc": "^12.0.2",
     "postcss-cssnext": "^3.1.0",
     "postcss-import": "^11.1.0",
-    "postcss-loader": "^2.1.4",
+    "postcss-loader": "^2.1.5",
     "postcss-url": "^7.3.2",
     "react-test-renderer": "^16.4.0",
     "recaptcha-test-keys": "^1.0.0",
     "rimraf": "^2.6.2",
-    "sinon": "^5.0.1",
-    "sinon-chai": "^3.0.0",
+    "sinon": "^5.1.1",
+    "sinon-chai": "^3.1.0",
     "style-loader": "^0.21.0",
     "through2": "^2.0.3",
     "u-wave-web-emojione": "^2.0.1",
     "uglifyjs-webpack-plugin": "^1.2.5",
-    "webpack": "^4.6.0",
-    "webpack-bundle-analyzer": "^2.11.2",
+    "webpack": "^4.12.0",
+    "webpack-bundle-analyzer": "^2.13.1",
     "webpack-dev-middleware": "^3.1.3",
-    "webpack-hot-middleware": "^2.22.1",
+    "webpack-hot-middleware": "^2.22.2",
     "webpack-merge": "^4.1.2",
     "webpack-pwa-manifest": "^3.6.2",
     "webpack-subresource-integrity": "^1.1.0-rc.4",
