@@ -12,6 +12,7 @@ const merge = require('webpack-merge');
 const htmlMinifierOptions = require('./tasks/utils/htmlMinifierOptions');
 
 const nodeEnv = process.env.NODE_ENV || 'development';
+const isDemo = process.env.DEMO === '1';
 
 // Compile src/ on the fly so we can use components etc. during build time.
 require('@babel/register').default({
@@ -60,6 +61,9 @@ const plugins = [
     filename: 'password-reset.html',
     title: 'Reset Password',
     minify: nodeEnv === 'production' ? htmlMinifierOptions : false,
+  }),
+  new DefinePlugin({
+    'process.env.FORCE_TOKEN': JSON.stringify(isDemo),
   }),
   new ProgressPlugin(),
   new LodashModuleReplacementPlugin({
@@ -110,7 +114,10 @@ const base = {
   context: path.join(__dirname, 'src'),
   entry: {
     bugsnag: './_wlk/bugsnag.js',
-    app: ['./app.js', './app.css'],
+    app: [
+      isDemo ? './demo.js' : './app.js',
+      './app.css',
+    ],
     passwordReset: ['./password-reset/app.js'],
   },
   mode: nodeEnv === 'production' ? 'production' : 'development',
